+++ fs/src/fs/org/jnode/fs/ntfs/security/SecurityUtils.java
@@ -55,10 +55,10 @@ public class SecurityUtils {
         byte[] authorityBuffer = new byte[6];
         structure.getData(offset + 2, authorityBuffer, 0, authorityBuffer.length);
         long authority = BigEndian.getUInt48(authorityBuffer, 0); // Why is this big endian??
-        List<Integer> subAuthorities = new ArrayList<Integer>();
+        List<Long> subAuthorities = new ArrayList<Long>();
 
         for (int i = 0; i < subAuthorityCount; i++) {
-            subAuthorities.add(structure.getInt32(offset + 8 + (4 * i)));
+            subAuthorities.add(structure.getUInt32(offset + 8 + (4 * i)));
         }
 
         return new SecurityIdentifier(authority, subAuthorities);
